backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
torch
transformers
pillow
pymupdf
langchain
langchain-community
langchain-groq
faiss-cpu
